# KAFKA-9729: avoid readLock in authorizer ACL lookups