@lbradstreet 2020-01-07
I'm also seeing some occasional ACL related test failures. It may be due to the way updates are performed. I'll have to circle back to this PR when I get a bit more time and figure those out.