@ijuma 2020-01-28
Do we actually need the underlying map to be sorted? We could probably get better performance if we relaxed that requirement.