@rajinisivaram 2020-01-28
@ijuma We store ACLs sorted by resource name in order to search for prefixed ACLs. I think we used to store them in a HashMap before prefixed ACLs were introduced.