@ijuma 2020-01-28
@rajinisivaram sure, but couldn't we simply have two maps?