@ijuma 2020-03-19
A related JIRA: https://issues.apache.org/jira/browse/KAFKA-9729