@omkreddy 2020-03-26
I ran existing Acl benchmark test, to make sure there is no performance degradation in authorize() method.  Not observed any perf degradation.

**Old** 
```
Benchmark                                (aclCount)  (resourceCount)  Mode  Cnt   Score   Error  Units
AclAuthorizerBenchmark.testAuthorizer             5             5000  avgt   15   0.240 ± 0.014  ms/op
AclAuthorizerBenchmark.testAuthorizer             5            10000  avgt   15   0.293 ± 0.020  ms/op
AclAuthorizerBenchmark.testAuthorizer             5            50000  avgt   15   0.566 ± 0.049  ms/op
AclAuthorizerBenchmark.testAuthorizer            10             5000  avgt   15   0.308 ± 0.057  ms/op
AclAuthorizerBenchmark.testAuthorizer            10            10000  avgt   15   0.271 ± 0.026  ms/op
AclAuthorizerBenchmark.testAuthorizer            10            50000  avgt   15   0.426 ± 0.026  ms/op
AclAuthorizerBenchmark.testAuthorizer            15             5000  avgt   15   0.236 ± 0.011  ms/op
AclAuthorizerBenchmark.testAuthorizer            15            10000  avgt   15   0.257 ± 0.012  ms/op
AclAuthorizerBenchmark.testAuthorizer            15            50000  avgt   15   0.508 ± 0.043  ms/op
```

**New**
```
Benchmark                              (aclCount)  (resourceCount)  Mode  Cnt  Score   Error  Units
AclAuthorizerBenchmark.testAuthorizer           5             5000  avgt   15  0.219 ± 0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer           5            10000  avgt   15  0.254 ± 0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer           5            50000  avgt   15  0.461 ± 0.002  ms/op
AclAuthorizerBenchmark.testAuthorizer          10             5000  avgt   15  0.229 ± 0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer          10            10000  avgt   15  0.235 ± 0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer          10            50000  avgt   15  0.462 ± 0.008  ms/op
AclAuthorizerBenchmark.testAuthorizer          15             5000  avgt   15  0.220 ± 0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer          15            10000  avgt   15  0.250 ± 0.001  ms/op
AclAuthorizerBenchmark.testAuthorizer          15            50000  avgt   15  0.422 ± 0.004  ms/op
JMH benchmarks done
```